Bundle binding advanced options section in component
